mediacapture-image/permission.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>

promise_test(async t => {
  const permissionDescriptor = { name: 'camera', panTiltZoom: true };
  await test_driver.set_permission(permissionDescriptor, 'granted', false);
 
  const permissionStatus = await navigator.permissions.query(permissionDescriptor);
  assert_equals(permissionStatus.state, 'granted');
}, 'test that we can set Camera PTZ permission in web platform test');

</script>
